:recycle: Massive refactor
