+++ src/learnatvcs.rs/src/lib.rs
@@ -1,4 +1,3 @@
-// TODO: optimize
 #![warn(missing_docs)]
 //! Fast request-based scraping API for learn@vcs
 
@@ -7,8 +6,11 @@ pub mod datematcher;
 pub use datematcher::Date;
 /// Error types
 pub mod errors;
+pub use crate::models::{TargetDate, TargetQuarter};
+mod models;
 mod scraper;
-pub use crate::scraper::{scrape, TargetDate, TargetQuarter};
+mod utils;
+pub use crate::scraper::scrape;
 #[cfg(test)]
 mod tests {
     // use super::*;
